cambios

+++ sistemaAvisosTSG/Models/TBCAS_AVISOS_ESTADO.cs
@@ -7,7 +7,7 @@ namespace sistemaAvisosTSG.Models
 {
     public class TBCAS_AVISOS_ESTADO
     {
-        public int AVISO_ESTADO_NRO { get; set; }
+        public decimal AVISO_ESTADO_NRO { get; set; }
 
         public string AVISO_ESTADO_DESCRIP { get; set; }
     }
